streamlit/requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
databricks-connect>=14.0.0
databricks-sdk>=0.20.0
PyArrow>=15.0.0
